Some more bits & pieces
- split worker into two helpers, moved functions from main
- separated units in csv export
- added test cases for split items
