+++ src/main/java/tf/gpx/edit/worker/GPXExtractCSVLinesWorker.java
@@ -41,7 +41,7 @@ import tf.gpx.edit.items.GPXWaypoint;
  * @author thomas
  */
 public class GPXExtractCSVLinesWorker extends GPXEmptyWorker {
-    private static final List<String> CSV_HEADER = Arrays.asList("ID", "Type", "Position", "Start/Date", "Name", "Duration", "Length [m]", "Speed [km/h]", "Height [m]");
+    private static final List<String> CSV_HEADER = Arrays.asList("ID", "Type", "Position", "Start/Date", "Name", "Duration", "Length", "LUnit", "Speed", "SUnit", "Height", "HUnit");
     
     private List<List<String>> csvLines = new ArrayList<>();
     
@@ -93,12 +93,14 @@ public class GPXExtractCSVLinesWorker extends GPXEmptyWorker {
     
     private List<String> visitGPXLineItem(final GPXLineItem gpxLineItem) {
         String length = "";
+        String lengthUnit = "m";
         String height = "";
         if (GPXLineItem.GPXLineItemType.GPXWaypoint.equals(gpxLineItem.getType())) {
             length = gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.DistanceToPrevious);
             height = gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.Elevation);
         } else {
-            length = gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.Length) + " [km]";
+            length = gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.Length);
+            lengthUnit = "km";
         }
         return Arrays.asList(
                 gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.CombinedID),
@@ -108,8 +110,11 @@ public class GPXExtractCSVLinesWorker extends GPXEmptyWorker {
                 gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.Name),
                 gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.Duration),
                 length,
+                lengthUnit,
                 gpxLineItem.getDataAsString(GPXLineItem.GPXLineItemData.Speed),
-                height
+                "km/h",
+                height,
+                "m"
             );
     }
 }
