+++ src/test/resources/testsplitmerge.gpx
